merge canvas stable

+++ app/controllers/quizzes/quizzes_controller.rb
@@ -28,12 +28,12 @@ class Quizzes::QuizzesController < ApplicationController
   # locked down.
   attr_reader :lock_results_if_needed
 
-  before_filter :require_context
-  before_filter :rich_content_service_config, only: [:show, :new, :edit]
+  before_action :require_context
+  before_action :rich_content_service_config, only: [:show, :new, :edit]
 
   add_crumb(proc { t('#crumbs.quizzes', "Quizzes") }) { |c| c.send :named_context_url, c.instance_variable_get("@context"), :context_quizzes_url }
-  before_filter { |c| c.active_tab = "quizzes" }
-  before_filter :require_quiz, :only => [
+  before_action { |c| c.active_tab = "quizzes" }
+  before_action :require_quiz, :only => [
     :statistics,
     :edit,
     :show,
@@ -47,8 +47,8 @@ class Quizzes::QuizzesController < ApplicationController
     :submission_html,
     :toggle_post_to_sis
   ]
-  before_filter :set_download_submission_dialog_title , only: [:show,:statistics]
-  after_filter :lock_results, only: [ :show, :submission_html ]
+  before_action :set_download_submission_dialog_title , only: [:show,:statistics]
+  after_action :lock_results, only: [ :show, :submission_html ]
   # The number of questions that can display "details". After this number, the "Show details" option is disabled
   # and the data is not even loaded.
   QUIZ_QUESTIONS_DETAIL_LIMIT = 25
@@ -104,8 +104,12 @@ class Quizzes::QuizzesController < ApplicationController
       skip_date_overrides: true,
       skip_lock_tests: true
     }]
+    max_name_length = AssignmentUtil.assignment_max_name_length(@context)
+    sis_name = AssignmentUtil.post_to_sis_friendly_name(@context)
+    due_date_required_for_account = AssignmentUtil.due_date_required_for_account?(@context)
+    sis_integration_settings_enabled = AssignmentUtil.sis_integration_settings_enabled?(@context)
 
-    js_env({
+    hash = {
       :QUIZZES => {
         assignment: quizzes_json(assignment_quizzes, *serializer_options),
         open: quizzes_json(open_quizzes, *serializer_options),
@@ -124,10 +128,21 @@ class Quizzes::QuizzesController < ApplicationController
       },
       :FLAGS => {
         question_banks: feature_enabled?(:question_banks),
-        post_to_sis_enabled: Assignment.sis_grade_export_enabled?(@context)
+        post_to_sis_enabled: Assignment.sis_grade_export_enabled?(@context),
+        migrate_quiz_enabled:  @domain_root_account.feature_enabled?(:quizzes2_exporter)
       },
       :quiz_menu_tools => external_tools_display_hashes(:quiz_menu),
-    })
+      :SIS_NAME => sis_name,
+      :MAX_NAME_LENGTH => max_name_length,
+      :DUE_DATE_REQUIRED_FOR_ACCOUNT => due_date_required_for_account,
+      :SIS_INTEGRATION_SETTINGS_ENABLED => sis_integration_settings_enabled
+    }
+    if @context.is_a?(Course) && @context.grants_right?(@current_user, session, :read)
+      hash[:COURSE_ID] = @context.id.to_s
+    end
+    js_env(hash)
+
+    set_tutorial_js_env
 
     conditional_release_js_env(includes: :active_rules)
 
@@ -226,6 +241,8 @@ class Quizzes::QuizzesController < ApplicationController
       js_env(hash)
       conditional_release_js_env(@quiz.assignment, includes: [:rule])
 
+      set_master_course_js_env_data(@quiz, @context)
+
       @quiz_menu_tools = external_tools_display_hashes(:quiz_menu)
       @can_take = can_take_quiz?
       if params[:take] && @can_take
@@ -258,6 +275,17 @@ class Quizzes::QuizzesController < ApplicationController
       if authorized_action(@quiz, @current_user, :update)
         @assignment = @quiz.assignment
       end
+
+      max_name_length_required_for_account = AssignmentUtil.name_length_required_for_account?(@context)
+      max_name_length = AssignmentUtil.assignment_max_name_length(@context)
+
+      hash = {
+        :MAX_NAME_LENGTH_REQUIRED_FOR_ACCOUNT => max_name_length_required_for_account,
+        :MAX_NAME_LENGTH => max_name_length,
+        :DUE_DATE_REQUIRED_FOR_ACCOUNT => AssignmentUtil.due_date_required_for_account?(@context),
+      }
+
+      js_env(hash)
       redirect_to(named_context_url(@context, :edit_context_quiz_url, @quiz))
     end
   end
@@ -266,7 +294,6 @@ class Quizzes::QuizzesController < ApplicationController
     if authorized_action(@quiz, @current_user, :update)
       add_crumb(@quiz.title, named_context_url(@context, :context_quiz_url, @quiz))
       @assignment = @quiz.assignment
-
       @quiz.title = params[:title] if params[:title]
       @quiz.due_at = params[:due_at] if params[:due_at]
       @quiz.assignment_group_id = params[:assignment_group_id] if params[:assignment_group_id]
@@ -283,11 +310,15 @@ class Quizzes::QuizzesController < ApplicationController
       end]
       sections = @context.course_sections.active
 
+      max_name_length_required_for_account = AssignmentUtil.name_length_required_for_account?(@context)
+      max_name_length = AssignmentUtil.assignment_max_name_length(@context)
+
       hash = {
         :ASSIGNMENT_ID => @assignment.present? ? @assignment.id : nil,
         :ASSIGNMENT_OVERRIDES => assignment_overrides_json(@quiz.overrides_for(@current_user,
                                                            ensure_set_not_empty: true),
                                                            @current_user),
+        :DUE_DATE_REQUIRED_FOR_ACCOUNT => AssignmentUtil.due_date_required_for_account?(@context),
         :QUIZ => quiz_json(@quiz, @context, @current_user, session),
         :SECTION_LIST => sections.map { |section|
           {
@@ -305,10 +336,12 @@ class Quizzes::QuizzesController < ApplicationController
         :quiz_max_combination_count => QUIZ_MAX_COMBINATION_COUNT,
         :SHOW_QUIZ_ALT_TEXT_WARNING => true,
         :VALID_DATE_RANGE => CourseDateRange.new(@context),
-        :MULTIPLE_GRADING_PERIODS_ENABLED => @context.feature_enabled?(:multiple_grading_periods)
+        :HAS_GRADING_PERIODS => @context.grading_periods?,
+        :MAX_NAME_LENGTH_REQUIRED_FOR_ACCOUNT => max_name_length_required_for_account,
+        :MAX_NAME_LENGTH => max_name_length
       }
 
-      if @context.feature_enabled?(:multiple_grading_periods)
+      if @context.grading_periods?
         hash[:active_grading_periods] = GradingPeriod.json_for(@context, @current_user)
       end
 
@@ -316,6 +349,7 @@ class Quizzes::QuizzesController < ApplicationController
       js_env(hash)
 
       conditional_release_js_env(@quiz.assignment)
+      set_master_course_js_env_data(@quiz, @context)
 
       render :new
     end
@@ -332,39 +366,39 @@ class Quizzes::QuizzesController < ApplicationController
         return render_forbidden unless grading_periods_allow_assignment_overrides_batch_create?(@quiz, overrides)
       end
 
-      params[:quiz][:title] = nil if params[:quiz][:title] == "undefined"
-      params[:quiz][:title] ||= t(:default_title, "New Quiz")
-      params[:quiz].delete(:points_possible) unless params[:quiz][:quiz_type] == 'graded_survey'
-      params[:quiz][:access_code] = nil if params[:quiz][:access_code] == ""
-      if params[:quiz][:quiz_type] == 'assignment' || params[:quiz][:quiz_type] == 'graded_survey'
-        params[:quiz][:assignment_group_id] ||= @context.assignment_groups.first.id
-        if (assignment_group_id = params[:quiz].delete(:assignment_group_id)) && assignment_group_id.present?
+      quiz_params = get_quiz_params
+      quiz_params[:title] = nil if quiz_params[:title] == "undefined"
+      quiz_params[:title] ||= t(:default_title, "New Quiz")
+      quiz_params.delete(:points_possible) unless quiz_params[:quiz_type] == 'graded_survey'
+      quiz_params[:access_code] = nil if quiz_params[:access_code] == ""
+      if quiz_params[:quiz_type] == 'assignment' || quiz_params[:quiz_type] == 'graded_survey'
+        quiz_params[:assignment_group_id] ||= @context.assignment_groups.first.id
+        if (assignment_group_id = quiz_params.delete(:assignment_group_id)) && assignment_group_id.present?
           @assignment_group = @context.assignment_groups.active.where(id: assignment_group_id).first
         end
         if @assignment_group
-          @assignment = @context.assignments.build(:title => params[:quiz][:title], :due_at => params[:quiz][:lock_at], :submission_types => 'online_quiz')
+          @assignment = @context.assignments.build(:title => quiz_params[:title], :due_at => quiz_params[:lock_at], :submission_types => 'online_quiz')
           @assignment.assignment_group = @assignment_group
           @assignment.saved_by = :quiz
           @assignment.workflow_state = 'unpublished'
           @assignment.save
-          params[:quiz][:assignment_id] = @assignment.id
+          quiz_params[:assignment_id] = @assignment.id
         end
-        params[:quiz][:assignment_id] = nil unless @assignment
-        params[:quiz][:title] = @assignment.title if @assignment
+        quiz_params[:assignment_id] = nil unless @assignment
+        quiz_params[:title] = @assignment.title if @assignment
       end
       @quiz.content_being_saved_by(@current_user)
       @quiz.infer_times
       @quiz.transaction do
-        @quiz.update_attributes!(params[:quiz])
+        @quiz.update_attributes!(quiz_params)
         batch_update_assignment_overrides(@quiz, overrides, @current_user) unless overrides.nil?
       end
-      if Assignment.sis_grade_export_enabled?(@context) && @quiz.assignment
-        post_to_sis = nil
-        post_to_sis = params[:assignment][:post_to_sis] if params[:assignment].present?
-        post_to_sis = @context.account.sis_default_grade_export[:value] if post_to_sis.nil?
-        @quiz.assignment.post_to_sis = post_to_sis
-        @quiz.assignment.save
+
+      if params[:post_to_sis]
+        @quiz.assignment.post_to_sis = params[:post_to_sis] == '1' ? true : false
       end
+
+
       @quiz.did_edit if @quiz.created?
       @quiz.reload
       render :json => @quiz.as_json(:include => {:assignment => {:include => :assignment_group}})
@@ -376,10 +410,11 @@ class Quizzes::QuizzesController < ApplicationController
   def update
     n = Time.now.to_f
     if authorized_action(@quiz, @current_user, :update)
+      quiz_params = get_quiz_params
       params[:quiz] ||= {}
 
       return render_forbidden unless grading_periods_allow_submittable_update?(
-        @quiz, params[:quiz], flash_message: true
+        @quiz, quiz_params, flash_message: true
       )
       overrides = delete_override_params
 
@@ -391,20 +426,20 @@ class Quizzes::QuizzesController < ApplicationController
         return render_forbidden unless batch_update_allowed
       end
 
-      params[:quiz][:title] = t("New Quiz") if params[:quiz][:title] == "undefined"
-      params[:quiz].delete(:points_possible) unless params[:quiz][:quiz_type] == 'graded_survey'
-      params[:quiz][:access_code] = nil if params[:quiz][:access_code] == ""
-      if params[:quiz][:quiz_type] == 'assignment' || params[:quiz][:quiz_type] == 'graded_survey' #'new' && params[:quiz][:assignment_group_id]
-        if (assignment_group_id = params[:quiz].delete(:assignment_group_id)) && assignment_group_id.present?
+      quiz_params[:title] = t("New Quiz") if quiz_params[:title] == "undefined"
+      quiz_params.delete(:points_possible) unless quiz_params[:quiz_type] == 'graded_survey'
+      quiz_params[:access_code] = nil if quiz_params[:access_code] == ""
+      if quiz_params[:quiz_type] == 'assignment' || quiz_params[:quiz_type] == 'graded_survey' #'new' && params[:quiz][:assignment_group_id]
+        if (assignment_group_id = quiz_params.delete(:assignment_group_id)) && assignment_group_id.present?
           @assignment_group = @context.assignment_groups.active.where(id: assignment_group_id).first
         end
         @assignment_group ||= @context.assignment_groups.first
         # The code to build an assignment for a quiz used to be here, but it's
         # been moved to the model quiz.rb instead.  See Quiz:build_assignment.
-        params[:quiz][:assignment_group_id] = @assignment_group && @assignment_group.id
+        quiz_params[:assignment_group_id] = @assignment_group && @assignment_group.id
       end
 
-      params[:quiz][:lock_at] = nil if params[:quiz].delete(:do_lock_at) == 'false'
+      quiz_params[:lock_at] = nil if quiz_params.delete(:do_lock_at) == 'false'
 
       @quiz.with_versioning(false) do
         @quiz.did_edit if @quiz.created?
@@ -414,16 +449,13 @@ class Quizzes::QuizzesController < ApplicationController
       respond_to do |format|
         @quiz.transaction do
           notify_of_update = value_to_boolean(params[:quiz][:notify_of_update])
-          params[:quiz][:notify_of_update] = false
 
           old_assignment = nil
           if @quiz.assignment.present?
             old_assignment = @quiz.assignment.clone
             old_assignment.id = @quiz.assignment.id
 
-            if params[:assignment] && Assignment.sis_grade_export_enabled?(@context)
-              @quiz.assignment.post_to_sis = params[:assignment][:post_to_sis]
-            end
+            @quiz.assignment.post_to_sis = params[:post_to_sis] == '1' ? true : false
           end
 
           auto_publish = @quiz.published?
@@ -431,7 +463,7 @@ class Quizzes::QuizzesController < ApplicationController
           @quiz.with_versioning(auto_publish) do
             # using attributes= here so we don't need to make an extra
             # database call to get the times right after save!
-            @quiz.attributes = params[:quiz]
+            @quiz.attributes = quiz_params
             @quiz.infer_times
             @quiz.content_being_saved_by(@current_user)
             if auto_publish
@@ -478,6 +510,7 @@ class Quizzes::QuizzesController < ApplicationController
 
   def destroy
     if authorized_action(@quiz, @current_user, :delete)
+      return render_unauthorized_action if editing_restricted?(@quiz)
       respond_to do |format|
         if @quiz.destroy
           format.html { redirect_to course_quizzes_url(@context) }
@@ -680,7 +713,7 @@ class Quizzes::QuizzesController < ApplicationController
     if authorized_action(@quiz, @current_user, :grade)
       @students = @context.students_visible_to(@current_user, include: :inactive)
       @students = @students.name_like(params[:search_term]) if params[:search_term].present?
-      @students = @students.uniq.order_by_sortable_name
+      @students = @students.distinct.order_by_sortable_name
       @students = @students.order(:uuid) if @quiz.survey? && @quiz.anonymous_submissions
       last_updated_at = Time.parse(params[:last_updated_at]) rescue nil
       respond_to do |format|
@@ -705,7 +738,7 @@ class Quizzes::QuizzesController < ApplicationController
       if @versions.size > 0 && !@quiz.muted?
         render :layout => false
       else
-        render :nothing => true
+        head :ok
       end
     end
   end
@@ -728,7 +761,7 @@ class Quizzes::QuizzesController < ApplicationController
       @lock_results_if_needed = true
       render layout: false
     else
-      render nothing: true
+      head :ok
     end
   end
 
@@ -958,5 +991,11 @@ class Quizzes::QuizzesController < ApplicationController
         format.json { render json: @quiz.errors, status: :forbidden }
       end
     end
+  end
+
+  protected
+
+  def get_quiz_params
+    params[:quiz] ? params[:quiz].permit(API_ALLOWED_QUIZ_INPUT_FIELDS[:only]) : {}
   end
 end
